Galaxy Parameter Classes

Galaxy Parameter Classes These classes encapsulate parameters to be passed to functions exposed in Galaxy.

GalaxyIntegerParam(...)

  GalaxyNumericParam(...)

  GalaxyCharacterParam(...)

  GalaxyLogicalParam(...)

  GalaxySelectParam(...)

Arguments

…
Arguments can be a single unnamed argument, corresponding to the name of the class, e.g., integer, numeric, character, or logical.

The first argument can be missing, can be of length 1 or can be a vector of a greater length, in which case it is rendered as a dropdown (select) list.

testValues:

(optional) Either an integer(1), numeric(1), character(1), or logical(1), depending on the class used. A value to be used in functional testing (running your function with a fixed set of inputs and checking that the outputs match what is expected).

Additional parameters are as follows:

label:

A character(1). The friendly name for this parameter. By default, RGalaxy will use the parameter's name ("fooBar" is changed to "Foo Bar") if this is not supplied.

min:

(optional) An integer(1). If the parameter is integer or numeric, this specifies a minimum value for the parameter. Galaxy will not allow lower values.

max:

(optional) An integer(1). If the parameter is integer or numeric, this specifies a maximum value for the parameter. Galaxy will not allow higher values.

force_select:

(default: FALSE) A logical(1). If this parameter is a dropdown list (i.e., the first argument is a vector of length > 1), a TRUE value forces the user to select an item.

display:

(optional) A character(1). If parameter is a dropdown list, this can be set to "checkboxes" or "radio" which renders the list as a set of check boxes or radio buttons. Defaults to a drop-down menu select list.

checked:

(default: FALSE) A logical(1). Whether the rendered checkbox should be checked. Only applicable if the parameter is a logical of length < 2.

size:

(default: 60L) An integer(1) The width of the field, in characters. For character parameters only.

required:

(default: FALSE) A logical(1). Whether Galaxy will require this field be filled in.

requiredMsg:

(default: "This field is required.") A character(1) If required is TRUE, the message to display to the user when the leave the field empty.

Details

The arguments to these classes are based on http://wiki.g2.bx.psu.edu/Admin/Tools/Tool%20Config%20Syntax#A.3Cparam.3E_tag_set

Examples

        GalaxyIntegerParam()

#> Found more than one class "GalaxyIntegerParam" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

#> Found more than one class "GalaxyIntegerParam" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

#> An object of class "GalaxyIntegerParam" (from package "RGalaxify")
#> integer(0)
#> Slot "testValues":
#> integer(0)
#> 
#> Slot "label":
#> character(0)
#> 
#> Slot "min":
#> numeric(0)
#> 
#> Slot "max":
#> numeric(0)
#> 
#> Slot "force_select":
#> [1] FALSE
#> 
#> Slot "display":
#> character(0)
#> 
#> Slot "checked":
#> [1] FALSE
#> 
#> Slot "size":
#> [1] 60
#> 
#> Slot "required":
#> [1] FALSE
#> 
#> Slot "requiredMsg":
#> [1] "This field is required."
#> 

        GalaxyIntegerParam(1L)

#> Found more than one class "MsgClass" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

#> Found more than one class "MsgClass" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

#> An object of class "GalaxyIntegerParam" (from package "RGalaxify")
#> [1] 1
#> Slot "testValues":
#> integer(0)
#> 
#> Slot "label":
#> character(0)
#> 
#> Slot "min":
#> numeric(0)
#> 
#> Slot "max":
#> numeric(0)
#> 
#> Slot "force_select":
#> [1] FALSE
#> 
#> Slot "display":
#> character(0)
#> 
#> Slot "checked":
#> [1] FALSE
#> 
#> Slot "size":
#> [1] 60
#> 
#> Slot "required":
#> [1] FALSE
#> 
#> Slot "requiredMsg":
#> [1] "This field is required."
#> 

        GalaxySelectParam(c(a=1L, b=2L))

#> Found more than one class "GalaxySelectParam" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

#> Found more than one class "GalaxySelectParam" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

#> An object of class "GalaxySelectParam" (from package "RGalaxify")
#> [1] 1 2
#> Slot "testValues":
#> NULL
#> 
#> Slot "label":
#> character(0)
#> 
#> Slot "min":
#> numeric(0)
#> 
#> Slot "max":
#> numeric(0)
#> 
#> Slot "force_select":
#> [1] FALSE
#> 
#> Slot "display":
#> character(0)
#> 
#> Slot "checked":
#> [1] FALSE
#> 
#> Slot "size":
#> [1] 60
#> 
#> Slot "required":
#> [1] FALSE
#> 
#> Slot "requiredMsg":
#> [1] "This field is required."
#> 

        GalaxyNumericParam(2.0, required=TRUE)

#> Found more than one class "GalaxyNumericParam" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

#> Found more than one class "GalaxyNumericParam" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

#> An object of class "GalaxyNumericParam" (from package "RGalaxify")
#> [1] 2
#> Slot "testValues":
#> numeric(0)
#> 
#> Slot "label":
#> character(0)
#> 
#> Slot "min":
#> numeric(0)
#> 
#> Slot "max":
#> numeric(0)
#> 
#> Slot "force_select":
#> [1] FALSE
#> 
#> Slot "display":
#> character(0)
#> 
#> Slot "checked":
#> [1] FALSE
#> 
#> Slot "size":
#> [1] 60
#> 
#> Slot "required":
#> [1] TRUE
#> 
#> Slot "requiredMsg":
#> [1] "This field is required."
#> 

        GalaxySelectParam(c("First Choice"="one",
                "Second Choice"="two")) # a select list

#> An object of class "GalaxySelectParam" (from package "RGalaxify")
#> [1] "one" "two"
#> Slot "testValues":
#> NULL
#> 
#> Slot "label":
#> character(0)
#> 
#> Slot "min":
#> numeric(0)
#> 
#> Slot "max":
#> numeric(0)
#> 
#> Slot "force_select":
#> [1] FALSE
#> 
#> Slot "display":
#> character(0)
#> 
#> Slot "checked":
#> [1] FALSE
#> 
#> Slot "size":
#> [1] 60
#> 
#> Slot "required":
#> [1] FALSE
#> 
#> Slot "requiredMsg":
#> [1] "This field is required."
#> 

        GalaxyLogicalParam(checked=TRUE)

#> Found more than one class "GalaxyLogicalParam" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

#> Found more than one class "GalaxyLogicalParam" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

#> An object of class "GalaxyLogicalParam" (from package "RGalaxify")
#> logical(0)
#> Slot "testValues":
#> logical(0)
#> 
#> Slot "label":
#> character(0)
#> 
#> Slot "min":
#> numeric(0)
#> 
#> Slot "max":
#> numeric(0)
#> 
#> Slot "force_select":
#> [1] FALSE
#> 
#> Slot "display":
#> character(0)
#> 
#> Slot "checked":
#> [1] TRUE
#> 
#> Slot "size":
#> [1] 60
#> 
#> Slot "required":
#> [1] FALSE
#> 
#> Slot "requiredMsg":
#> [1] "This field is required."
#>