Class "GalaxyInputFile"

Represents a dataset hosted on Galaxy, to be passed to an R function that you expose in Galaxy.

GalaxyInputFile(required=TRUE, formatFilter=character(0))

Arguments

required
Whether to require that the user provide this dataset.

formatFilter
(optional) A file type for filtering the list of possible input data sets. Should be one of the file types listed in the datatypes_conf.xml file in the root of your Galaxy distribution.

Methods

No methods defined with class "GalaxyInputFile" in the signature.

See also

link{galaxy}, GalaxyConfig

Examples

    inputfile1=GalaxyInputFile()

#> Found more than one class "GalaxyInputFile" in cache; using the first, from namespace 'RGalaxy'

#> Also defined by ‘RGalaxify’

    ## The user can only choose csv files:
    inputfile2=GalaxyInputFile(formatFilter="csv")