Add two numbers (with functional test)

An example function that can be made into a Galaxy tool. Takes two numbers, adds them, and returns a file containing the result. This version demonstrates a functional test.

addTwoNumbersWithTest(number1=GalaxyNumericParam(required=TRUE, testValues=5L),
        number2=GalaxyNumericParam(required=TRUE, testValues=5L),
        sum=GalaxyOutput("sum", "txt"))

Arguments

number1
The first number to add.

number2
The second number to add.

sum
Where the result of the addition should be written.

Value

invisible(NULL)

See also

galaxy, GalaxyConfig, GalaxyOutput, addTwoNumbers

Examples

t <- tempfile()
addTwoNumbersWithTest(2, 2, t)
readLines(t, warn=FALSE)

#> [1] "4"

runFunctionalTest(addTwoNumbersWithTest) ## should return TRUE

#> Error in file.path(pkg_path, "inst", ...): argument "funcName" is missing, with no default