Sends informational, warning, and error messages to the user.

Send error, warning, and informational messages to the user. Use these instead of message, warning, and stop. Output is wrapped consistently and passed through sprintf so you can use inline formatting (see examples). Output of gstop will appear in Galaxy user's web browser.

gmessage(..., appendLF = TRUE)

  gstop(..., call. = FALSE)

  gwarning(..., call. = FALSE, immediate. = FALSE)

Arguments

...
Passed to sprintf.

appendLF
Passed to message.

call.
Passed to stop or warning.

immediate.
Passed to warning.

Value

NULL

See also

message, warning, stop, sprintf

Examples

gmessage("This is an %s message.", "example")

#> This is an example message.
# NOT RUN {
gstop("Encountered a %s error.", "serious")
# }# NOT RUN {
 gwarning("Something is not quite right.")
# }